add purge/force/recurse to $config_path/pdns.d
It seems (at least) on Puppet 4.7/4.8 that defining a directory file
resource (directory B) as a child in the directory (directory A) being
recursed/purged, that it prevents children of directory B being purged.

The work around for this is to simply include the purge/force/recurse flags
to the pdns.d directory as well.